Change metadata and use dark theme by default
Change out the metadata like the author name to my name and also show
instead of the twitter handle, the email address from metadata. As
mentioned also use dark theme by default.

This would be just a template for my presentations so I won't have to
change all that whenever I wan't to use it. All the credit for making
this template go to the original author.